reports/requirements.txt
fastapi[standard]
requests
PyJWT
cryptography
python-decouple
